@@ -1,11 +1,8 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
-import './assets/doc.scss'
+import '@/assets/doc.scss'
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-  router
-}).$mount('#app')
+createApp(App)
+.use(router)
+.mount('#app')
